Refactor console and fix dedicated on windows
